fix a few errors found by findbugs

+++ org/gjt/sp/jedit/BufferHistory.java
@@ -34,7 +34,7 @@ import org.gjt.sp.util.Log;
 /**
  * Recent file list.
  * @author Slava Pestov
- * @version $Id: BufferHistory.java,v 1.14 2004/03/19 19:16:35 spestov Exp $
+ * @version $Id: BufferHistory.java,v 1.15 2004/05/29 01:55:24 spestov Exp $
  */
 public class BufferHistory
 {
@@ -171,10 +171,13 @@ public class BufferHistory
 
 		String lineSep = System.getProperty("line.separator");
 
+		boolean ok = false;
+
+		BufferedWriter out = null;
+
 		try
 		{
-			BufferedWriter out = new BufferedWriter(
-				new FileWriter(file1));
+			out = new BufferedWriter(new FileWriter(file1));
 
 			out.write("<?xml version=\"1.0\"?>");
 			out.write(lineSep);
@@ -227,14 +230,30 @@ public class BufferHistory
 
 			out.close();
 
-			/* to avoid data loss, only do this if the above
-			 * completed successfully */
-			file2.delete();
-			file1.renameTo(file2);
+			ok = true;
 		}
 		catch(Exception e)
 		{
 			Log.log(Log.ERROR,BufferHistory.class,e);
+		}
+		finally
+		{
+			try
+			{
+				if(out != null)
+					out.close();
+			}
+			catch(IOException e)
+			{
+			}
+		}
+
+		if(ok)
+		{
+			/* to avoid data loss, only do this if the above
+			 * completed successfully */
+			file2.delete();
+			file1.renameTo(file2);
 		}
 
 		recentModTime = file2.lastModified();
